# src/main/resources/openapi/graph-api-v1.yaml
openapi: 3.1.0
info:
  title: Investigation Graph API
  version: 1.0.0
  description: Investigation graph API with switchable graph-query backend, smart 1-hop expansion and bounded export.
servers:
  - url: /api/v1
paths:
  /graph/expand:
    post:
      summary: Smart 1-hop expansion for investigation graph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphExpandRequest'
      responses:
        '200':
          description: Expanded graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphExpandResponse'
        '503':
          description: DuckPGQ is required but unavailable
  /graph/shortest-path:
    post:
      summary: Find shortest path between two nodes inside relation family
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortestPathRequest'
      responses:
        '200':
          description: Shortest path graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortestPathResponse'
        '404':
          description: Path not found
        '503':
          description: DuckPGQ is required but unavailable
  /graph/dictionary:
    get:
      summary: Get edge/status dictionary for UI legend
      responses:
        '200':
          description: Dictionary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDictionaryResponse'
  /graph/node-summary:
    get:
      summary: Get node click summary before graph expansion
      parameters:
        - name: nodeId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: relationFamily
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/GraphRelationFamily'
        - name: direction
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Direction'
            default: BOTH
      responses:
        '200':
          description: Node summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphNodeSummaryResponse'
        '404':
          description: Node not found
  /graph/nodes/search:
    get:
      summary: Search graph nodes by name, identifier or attribute text
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
        - name: nodeType
          in: query
          required: false
          schema:
            type: string
            description: Optional node type filter, e.g. PERSON, ACCOUNT or COMPANY.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: includeAttributes
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Matching graph nodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphNodeSearchResponse'
        '400':
          description: Invalid search query or limit
  /graph/export:
    post:
      summary: Export rendered graph as JSON, CSV or NDJSON
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [JSON, CSV, NDJSON]
            default: JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphExportRequest'
      responses:
        '200':
          description: File payload
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
            application/x-ndjson:
              schema:
                type: string
components:
  schemas:
    GraphExpandRequest:
      type: object
      required: [seeds, direction]
      properties:
        seeds:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/SeedRef'
        relationFamily:
          $ref: '#/components/schemas/GraphRelationFamily'
        edgeTypes:
          type: array
          maxItems: 100
          items:
            type: string
        direction:
          $ref: '#/components/schemas/Direction'
        maxNeighborsPerSeed:
          type: integer
          minimum: 1
          maximum: 1000
        maxNodes:
          type: integer
          minimum: 1
          maximum: 1000
        maxEdges:
          type: integer
          minimum: 1
          maximum: 5000
        includeAttributes:
          type: boolean
    GraphExpandResponse:
      type: object
      required: [nodes, edges, meta]
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'
        meta:
          $ref: '#/components/schemas/GraphMeta'
    ShortestPathRequest:
      type: object
      required: [source, target, direction]
      properties:
        source:
          $ref: '#/components/schemas/SeedRef'
        target:
          $ref: '#/components/schemas/SeedRef'
        relationFamily:
          $ref: '#/components/schemas/GraphRelationFamily'
        direction:
          $ref: '#/components/schemas/Direction'
        maxDepth:
          type: integer
          minimum: 1
          maximum: 10
    ShortestPathResponse:
      type: object
      required: [path, nodes, edges, meta]
      properties:
        path:
          $ref: '#/components/schemas/Path'
        nodes:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          maxItems: 5000
          items:
            $ref: '#/components/schemas/GraphEdge'
        meta:
          $ref: '#/components/schemas/GraphMeta'
    GraphExportRequest:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          maxItems: 5000
          items:
            $ref: '#/components/schemas/GraphEdge'
        meta:
          $ref: '#/components/schemas/GraphMeta'
    GraphNodeSummaryResponse:
      type: object
      required: [node, summary, relationFamilies, edgeTypes, neighborNodeTypes, expandPreview]
      properties:
        node:
          $ref: '#/components/schemas/GraphNode'
        summary:
          $ref: '#/components/schemas/GraphNodeSummary'
        relationFamilies:
          type: array
          items:
            $ref: '#/components/schemas/GraphFacetCount'
        edgeTypes:
          type: array
          items:
            $ref: '#/components/schemas/GraphFacetCount'
        neighborNodeTypes:
          type: array
          items:
            $ref: '#/components/schemas/GraphFacetCount'
        expandPreview:
          $ref: '#/components/schemas/GraphExpandPreview'
    GraphNodeSearchResponse:
      type: object
      required: [nodes, meta]
      properties:
        nodes:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/GraphNode'
        meta:
          $ref: '#/components/schemas/GraphNodeSearchMeta'
    GraphNodeSearchMeta:
      type: object
      required: [query, limit, returnedNodeCount, truncated]
      properties:
        query:
          type: string
        nodeType:
          type:
            - string
            - 'null'
        limit:
          type: integer
        returnedNodeCount:
          type: integer
        truncated:
          type: boolean
    GraphDictionaryResponse:
      type: object
      required: [edgeTypes, relationFamilies, nodeTypes, nodeStatuses, styleHints]
      properties:
        edgeTypes:
          type: array
          items:
            type: string
        relationFamilies:
          type: array
          items:
            type: string
        nodeTypes:
          type: array
          items:
            type: string
        nodeStatuses:
          type: array
          items:
            type: string
        styleHints:
          type: object
          additionalProperties:
            type: string
    GraphNodeSummary:
      type: object
      required: [requestedDirection, relationFamily, adjacentEdgeCount, uniqueNeighborCount, outboundEdgeCount, inboundEdgeCount]
      properties:
        requestedDirection:
          $ref: '#/components/schemas/Direction'
        relationFamily:
          type: string
        adjacentEdgeCount:
          type: integer
        uniqueNeighborCount:
          type: integer
        outboundEdgeCount:
          type: integer
        inboundEdgeCount:
          type: integer
    GraphFacetCount:
      type: object
      required: [key, count]
      properties:
        key:
          type: string
        count:
          type: integer
    GraphExpandPreview:
      type: object
      required: [defaultMaxNeighborsPerSeed, defaultMaxNodes, defaultMaxEdges, wouldTruncateByNeighborBudget]
      properties:
        defaultMaxNeighborsPerSeed:
          type: integer
        defaultMaxNodes:
          type: integer
        defaultMaxEdges:
          type: integer
        wouldTruncateByNeighborBudget:
          type: boolean
    SeedRef:
      type: object
      required: [type, value]
      properties:
        type:
          $ref: '#/components/schemas/SeedType'
        value:
          type: string
          minLength: 1
    SeedType:
      type: string
      description: Identifier type stored in `g_identifiers`; aliases like `id`, `party`, or `phone` are accepted by the API.
    GraphRelationFamily:
      type: string
      description: Relation family value stored in `g_edges.relation_family`; `ALL_RELATIONS` disables family filtering.
    Direction:
      type: string
      enum: [OUTBOUND, INBOUND, BOTH]
    GraphNode:
      type: object
      required: [nodeId, displayName]
      properties:
        nodeId:
          type: string
          minLength: 1
        nodeType:
          type: string
        displayName:
          type: string
          minLength: 1
        identifiers:
          type: object
          additionalProperties:
            type: string
        statuses:
          type: array
          items:
            type: string
        attributes:
          type: object
          additionalProperties: true
    GraphEdge:
      type: object
      required: [edgeId, fromNodeId, toNodeId, type, directed]
      properties:
        edgeId:
          type: string
          minLength: 1
        fromNodeId:
          type: string
          minLength: 1
        toNodeId:
          type: string
          minLength: 1
        type:
          type: string
          minLength: 1
        relationFamily:
          type: string
        directed:
          type: boolean
        weight:
          type: number
        sourceSystem:
          type: string
        firstSeenAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
        attributes:
          type: object
          additionalProperties: true
    GraphMeta:
      type: object
      required: [truncated, executionMs, source, relationFamily, rankingStrategy, candidateEdgeCount, returnedNodeCount, returnedEdgeCount, warnings]
      properties:
        truncated:
          type: boolean
        executionMs:
          type: integer
        source:
          type: string
          enum: [DUCKPGQ, NEO4J]
        relationFamily:
          $ref: '#/components/schemas/GraphRelationFamily'
        rankingStrategy:
          type: string
        candidateEdgeCount:
          type: integer
        returnedNodeCount:
          type: integer
        returnedEdgeCount:
          type: integer
        warnings:
          type: array
          items:
            type: string
    Path:
      type: object
      required: [orderedNodeIds, orderedEdgeIds, hopCount]
      properties:
        orderedNodeIds:
          type: array
          items:
            type: string
        orderedEdgeIds:
          type: array
          items:
            type: string
        hopCount:
          type: integer